+++ tools/time.h
@@ -226,20 +226,20 @@ template<>
 struct std::formatter<evo::time::Nanoseconds> : std::formatter<int64_t> {
     auto format(const evo::time::Nanoseconds& time, std::format_context& ctx) {
         return std::formatter<int64_t>::format(evo::i64(time), ctx);
-    }
+    };
 };
 
 template<>
 struct std::formatter<evo::time::Milliseconds> : std::formatter<int64_t> {
     auto format(const evo::time::Milliseconds& time, std::format_context& ctx) {
         return std::formatter<int64_t>::format(evo::i64(time), ctx);
-    }
+    };
 };
 
 template<>
 struct std::formatter<evo::time::Seconds> : std::formatter<int64_t> {
     auto format(const evo::time::Seconds& time, std::format_context& ctx) {
         return std::formatter<int64_t>::format(evo::i64(time), ctx);
-    }
+    };
 };